followed coding conventions in naming CSS classes, corrected the path for the test, renamed landing-page directory to landing and removed the navbar
