+++ client/src/assets/Landing.css
@@ -10,14 +10,14 @@
 /*======================================
             Main Section
 ======================================*/
-.welcome_text {
+.welcome-text {
   margin-top: 15%;
   padding: 5%;
   line-height: 1.5em;
   letter-spacing: 0.3em;
 }
 
-.welcome_text_button {
+.welcome-text-button {
   padding: 10px;
   width: 20%;
   margin: 40px auto;
@@ -25,7 +25,7 @@
 /*======================================
             Section one
 ======================================*/
-.featuresOne {
+.features-one {
   margin-top: 70px;
   padding: 50px;
   letter-spacing: 0.1em;
@@ -34,7 +34,7 @@
 /*======================================
             Section two
 ======================================*/
-.featuresTwo {
+.features-two {
   margin-top: 170px;
   padding: 80px;
 }
@@ -42,7 +42,7 @@
 /*======================================
             Section three
 ======================================*/
-.featuresThree {
+.features-three {
   /* margin-top: 50px; */
   padding: 80px;
 }
