+++ client/src/components/landing/SectionThree.js
@@ -5,9 +5,9 @@ import dataTwo from '../../assets/images/dataTwo.jpg';
 
 const SectionThree = () => {
   return (
-    <section className="section_three section">
+    <section className="section-three section">
       <h3 className="text-center font-weight-bold mt-10"> Read easy to interpret, live data from all your devices </h3>
-      <div className="row featuresThree">
+      <div className="row features-three">
         <div className="col-md-6">
           <img src={dataOne} className="img-fluid rounded" alt="Internet of things devices" height="500px" width="500px" />
         </div>
