+++ client/src/tests/Landing.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Landing from '../Landing';
+import Landing from '../containers/Landing';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
